+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - CLI command build: deck author is now KRCG
 - Improve README.md
+- Add the krcg-gen tool, to generate static files for third parties
 
 0.23 (2020-04-24)
 -----------------